[J:MEP-0] Move ImageLoader install into ComposeInitializer + relocate path helper
Address review feedback on the desktop image-loader wiring:

- Replace the `installSnagDesktopImageLoader()` call from `main()` with
  `SnagDesktopImageLoaderInitializer : ComposeInitializer`. App.kt
  already iterates `getKoin().getAll<ComposeInitializer>()` synchronously
  before any child composes, so this is the right hook to register the
  Coil singleton before the first AsyncImage. Bound on JVM via the
  per-platform `platformModule` actual.

- Split the `nonWebMain` `platformModule` actual into per-platform
  actuals (`jvmMain`, `androidMain`, `iosMain`), matching the pattern
  webMain already uses. Only JVM contributes a non-empty module today.

- Move `resolveJvmAppDataDir(...)` from `core/foundation/fe` to
  `lib/storage/fe/api`. Filesystem layout is infrastructure knowledge,
  not a feature-agnostic core utility — `lib/storage` is the better
  home alongside the rest of the desktop file-storage code. Database
  module gets an explicit dependency on `lib.storage.fe.api`.
