+++ composeApp/build.gradle.kts
@@ -61,6 +61,7 @@ kotlin {
         jvmMain.dependencies {
             implementation(compose.desktop.currentOs)
             implementation(projects.lib.configuration.fe.api)
+            implementation(projects.lib.storage.fe.api)
             implementation(libs.ktor.client.okhttp)
         }
         val webMain by getting {
