+++ composeApp/src/iosMain/kotlin/cz/adamec/timotej/snag/di/AppModule.ios.kt
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2026 Timotej Adamec
+ * SPDX-License-Identifier: MIT
+ *
+ * This file is part of the thesis:
+ * "Multiplatform snagging system with code sharing maximisation"
+ *
+ * Czech Technical University in Prague
+ * Faculty of Information Technology
+ * Department of Software Engineering
+ */
+
+package cz.adamec.timotej.snag.di
+
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+internal actual val platformModule: Module = module {}
