+++ composeApp/src/jvmMain/kotlin/cz/adamec/timotej/snag/initializer/SnagDesktopImageLoaderInitializer.kt
@@ -10,29 +10,36 @@
  * Department of Software Engineering
  */
 
-package cz.adamec.timotej.snag
+package cz.adamec.timotej.snag.initializer
 
+import androidx.compose.runtime.Composable
 import coil3.ImageLoader
 import coil3.PlatformContext
 import coil3.SingletonImageLoader
+import coil3.annotation.ExperimentalCoilApi
 import coil3.disk.DiskCache
 import coil3.disk.directory
 import coil3.network.ktor3.KtorNetworkFetcherFactory
 import coil3.request.crossfade
 import coil3.util.DebugLogger
 import cz.adamec.timotej.snag.configuration.fe.FrontendRunConfig
-import cz.adamec.timotej.snag.core.foundation.fe.resolveJvmAppDataDir
+import cz.adamec.timotej.snag.lib.design.fe.initializer.ComposeInitializer
+import cz.adamec.timotej.snag.lib.storage.fe.api.resolveJvmAppDataDir
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.okhttp.OkHttp
 import okio.Path.Companion.toOkioPath
 import java.io.File
 
-internal fun installSnagDesktopImageLoader() {
-    SingletonImageLoader.setSafe { context ->
-        createSnagDesktopImageLoader(context = context)
+internal class SnagDesktopImageLoaderInitializer : ComposeInitializer {
+    @Composable
+    override fun init() {
+        SingletonImageLoader.setSafe { context ->
+            createSnagDesktopImageLoader(context = context)
+        }
     }
 }
 
+@OptIn(ExperimentalCoilApi::class)
 private fun createSnagDesktopImageLoader(context: PlatformContext): ImageLoader {
     val baseDir =
         resolveJvmAppDataDir(
