+++ lib/storage/fe/api/src/jvmTest/kotlin/cz/adamec/timotej/snag/lib/storage/fe/api/JvmAppDataDirTest.kt
@@ -10,7 +10,7 @@
  * Department of Software Engineering
  */
 
-package cz.adamec.timotej.snag.core.foundation.fe
+package cz.adamec.timotej.snag.lib.storage.fe.api
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
